requirements.txt
# === Agent + LLM Tools ===
aisuite[openai]==0.1.11
openai
docstring-parser
python-dotenv

# === Web Framework + API ===
fastapi
pydantic
pydantic[email]
requests
sqlalchemy
uvicorn

# === Notebook ===
ipykernel